+++ R/conda.R
@@ -80,7 +80,7 @@
 #' To force `reticulate` to use a particular `conda` binary, we recommend
 #' setting:
 #'
-#' ```
+#' ```r
 #' options(reticulate.conda_binary = "/path/to/conda")
 #' ```
 #'
@@ -238,7 +238,7 @@ conda_create <- function(envname = NULL,
     args <- c(args, "-c", ch)
 
   # invoke conda
-  result <- system2t(conda, shQuote(args))
+  result <- system2t(conda, maybe_shQuote(args))
   if (result != 0L) {
     fmt <- "Error creating conda environment '%s' [exit code %i]"
     stopf(fmt, envname, result, call. = FALSE)
@@ -260,10 +260,10 @@ conda_create_env <- function(envname, environment, conda) {
     if (is.null(envname))
       c()
     else if (grepl("/", envname))
-      c("--prefix", shQuote(envname))
+      c("--prefix", maybe_shQuote(envname))
     else
-      c("--name", shQuote(envname)),
-    "-f", shQuote(environment)
+      c("--name", maybe_shQuote(envname)),
+    "-f", maybe_shQuote(environment)
   )
 
   result <- system2t(conda, args)
@@ -302,7 +302,7 @@ conda_clone <- function(envname, ..., clone = "base", conda = "auto") {
   args <- c(args, "--clone", clone)
 
   # invoke conda
-  result <- system2t(conda, shQuote(args))
+  result <- system2t(conda, maybe_shQuote(args))
   if (result != 0L) {
     fmt <- "Error creating conda environment '%s' [exit code %i]"
     stopf(fmt, envname, result, call. = FALSE)
@@ -376,7 +376,7 @@ conda_remove <- function(envname,
 
   # remove packages (or the entire environment)
   args <- conda_args("remove", envname, packages)
-  result <- system2t(conda, shQuote(args))
+  result <- system2t(conda, maybe_shQuote(args))
   if (result != 0L) {
     stop("Error ", result, " occurred removing conda environment ", envname,
          call. = FALSE)
@@ -457,7 +457,7 @@ conda_install <- function(envname = NULL,
   # (should be no-op if that copy of Python already installed)
   if (!is.null(python_version)) {
     args <- conda_args("install", envname, python_package)
-    status <- system2t(conda, shQuote(args))
+    status <- system2t(conda, maybe_shQuote(args))
     if (status != 0L) {
       fmt <- "installation of '%s' into environment '%s' failed [error code %i]"
       msg <- sprintf(fmt, python_package, envname, status)
@@ -494,7 +494,7 @@ conda_install <- function(envname = NULL,
     args <- c(args, "-c", ch)
 
   args <- c(args, python_package, packages)
-  result <- system2t(conda, shQuote(args))
+  result <- system2t(conda, maybe_shQuote(args))
 
   # check for errors
   if (result != 0L) {
@@ -511,7 +511,7 @@ conda_install <- function(envname = NULL,
 conda_binary <- function(conda = "auto") {
 
   # automatic lookup if requested
-  if (identical(conda, "auto")) {
+  if (identical(conda, "auto") || isTRUE(is.na(conda))) {
     conda <- find_conda()
     if (is.null(conda))
       stop("Unable to find conda binary. Is Anaconda installed?", call. = FALSE)
@@ -581,7 +581,7 @@ conda_update <- function(conda = "auto") {
   name <- if ("anaconda" %in% envlist$name) "anaconda" else "conda"
 
   # attempt update
-  system2t(conda, c("update", "--prefix", shQuote(prefix), "--yes", name))
+  system2t(conda, c("update", "--prefix", maybe_shQuote(prefix), "--yes", name))
 }
 
 numeric_conda_version <- function(conda = "auto", version_string = conda_version(conda)) {
@@ -1048,8 +1048,12 @@ get_python_conda_info <- function(python) {
     conda <- python_info_condaenv_find(root)
   }
 
+  conda <- normalizePath(conda, winslash = "/", mustWork = FALSE)
+  if(!file.exists(conda))
+    conda <- NA
+
   list(
-    conda = normalizePath(conda, winslash = "/", mustWork = TRUE),
+    conda = conda,
     root = normalizePath(root, winslash = "/", mustWork = TRUE)
   )