+++ R/miniconda.R
@@ -239,14 +239,14 @@ miniconda_installer_run <- function(installer, update, path) {
   }
   if (is_windows()) {
     installer <- normalizePath(installer)
-    status <- system2(installer, args)
+    status <- system2t(installer, args)
   }
   if (is_unix()) {
     ##check for bash
     bash_path <- Sys.which("bash")
     if (bash_path[1] == "")
       stopf("The miniconda installer requires bash.")
-    status <- system2(bash_path[1], c(installer, args))
+    status <- system2t(bash_path[1], c(installer, args))
   }
   if (status != 0)
     stopf("miniconda installation failed [exit code %i]", status)
@@ -403,7 +403,7 @@ miniconda_python_envpath <- function() {
 
 # the version of python to use in the environment
 miniconda_python_version <- function() {
-  Sys.getenv("RETICULATE_MINICONDA_PYTHON_VERSION", unset = "3.9")
+  Sys.getenv("RETICULATE_MINICONDA_PYTHON_VERSION", unset = "3.10")
 }
 
 miniconda_python_package <- function() {